removing SampleDocument class
